#43 resolve issue, refactor in separate component
